Clean up example app
