+++ Example/Pageboy-Example/PageViewController+Appearance.swift
@@ -10,9 +10,8 @@ import UIKit
 
 extension PageViewController {
     
-    //
+    
     // MARK: Bar buttons
-    //
     
     func addBarButtons() {
         
@@ -31,11 +30,10 @@ extension PageViewController {
         self.nextBarButton?.isEnabled = index != (self.pageCount ?? 0) - 1
     }
     
-    //
+
     // MARK: Appearance
-    //
-    
-    func updateAppearance(pageOffset: CGFloat) {
+
+    func updateGradient(for pageOffset: CGFloat) {
         
         var offset = pageOffset
         if offset < 0.0 {
